+++ pubspec.yaml
@@ -17,6 +17,7 @@ dependencies:
   flutter_localization: ^0.2.1
   cupertino_icons: ^1.0.6
   loading_animation_widget: 1.2.1
+  socket_io_client: ^2.0.3+1
 
 dev_dependencies:
   flutter_test:
